requirements.txt
asgiref==3.9.1
Django==5.2.4
django-debug-toolbar==5.2.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
